Move version command to its own file and update readme
